requirements.txt
flask
Flask-Login
Flask-SQLAlchemy
google-generativeai
python-dotenv
librosa
matplotlib
numpy
werkzeug
gunicorn